old/pn532_reader/pyproject.toml
[tool.poetry]
name = "rfid-reader"
version = "0.1.0"
description = ""
authors = ["Max Maton <me@mmaton.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9 <=3.12"
pyftdi = "^0.55.0"
adafruit-blinka = "^8.22.4"
nfcpy = "^1.0.4"
tlv8 = "^0.10.0"
cryptography = "^41.0.4"
environs = "^9.5.0"


[tool.poetry.group.dev.dependencies]
ipython = "^8.16.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
